Add GeoScore computation and refactor training script
