+++ docs/api/autotune.md
@@ -1,0 +1,8 @@
+# model.autotune
+
+Optuna-based autotuning helpers used by `train.py`.
+
+::: model.autotune
+    options:
+      show_root_heading: false
+      members_order: source
