+++ docs/api/metrics.md
@@ -1,0 +1,8 @@
+# model.metrics
+
+Validation metric helpers used by training and evaluation.
+
+::: model.metrics
+    options:
+      show_root_heading: false
+      members_order: source
